ajustes
Former-commit-id: 363fa5c58382bdd66eff3d2dc607add64293c64e

+++ app/Console/Commands/ImportShopifyTracking.php
@@ -45,7 +45,7 @@ class ImportShopifyTracking extends Command
         $productService = new ProductService();
         $trackingService = new TrackingService();
 
-        $userToVerify = 147;
+        $userToVerify = 979;
 
         $count = 1;
 
@@ -55,7 +55,7 @@ class ImportShopifyTracking extends Command
             'productsPlansSale',
             'plansSales.plan.project.shopifyIntegrations'
         ])->doesntHave('tracking')
-            //->where('owner_id', $userToVerify)
+            ->where('owner_id', $userToVerify)
             ->where('status', 1)
             ->whereNotNull('shopify_order')
             ->orderByDesc('id')
@@ -74,7 +74,7 @@ class ImportShopifyTracking extends Command
                         $shopifyService = new ShopifyService($integration->url_store, $integration->token, false);
 
                         //Recria o webhook
-                        /*if(!in_array($project->id, $recreatedWebhooks)){
+                        if(!in_array($project->id, $recreatedWebhooks)){
 
                             $shopifyService->deleteShopWebhook();
 
@@ -99,7 +99,7 @@ class ImportShopifyTracking extends Command
                             $this->info('Webhook do projeto recriado');
 
                             $recreatedWebhooks[] = $project->id;
-                        }*/
+                        }
 
                         //Verifica os códigos de rastreio que não vieram nos postbacks
                         $fulfillments = $shopifyService->findFulfillments($sale->shopify_order);
